requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
rank-bm25==0.2.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
rich==13.8.0
PyYAML==6.0.2
jsonschema==4.23.0
PyJWT==2.8.0
python-multipart

# NEW for parsing/OCR/extraction/graphs
PyMuPDF==1.24.9           # fitz
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
python-docx==1.1.2
langdetect==1.0.9
dateparser==1.2.0
Babel==2.15.0
networkx==3.3
matplotlib==3.9.2

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
rank-bm25==0.2.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
rich==13.8.0
PyYAML==6.0.2
jsonschema==4.23.0
PyMuPDF==1.24.9
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
python-docx==1.1.2
langdetect==1.0.9
dateparser==1.2.0
Babel==2.15.0
networkx==3.3
matplotlib==3.9.2

# NEW for LoRA training & inference
torch==2.3.1
transformers==4.44.2
accelerate==0.33.0
peft==0.12.0

# NEW for Exports
reportlab==4.2.2
pyarrow==17.0.0

# NEW for UI + HTTP
streamlit==1.37.1
requests==2.32.3